Added confimation alert to delete button

+++ App-Tracker/src/Components/Delete.jsx
@@ -2,15 +2,19 @@ const Delete = ({ app }) => {
 
     //delete application based on jobapp_id
     const deleteApp = async (id) => {
-        try {
+        const confirmDelete = window.confirm('Do you want to delete');
+        if(confirmDelete === true)
+        {
+            try {
             const response = await fetch(`http://localhost:5000/apps/${id}`, {
                 method: 'DELETE',
                 headers: {'Content-type': 'application/json'},
                 body: JSON.stringify(app)
             });
-        window.location = '/';
-        } catch (error) {
-            console.error(error.message);
+            window.location = '/';
+            } catch (error) {
+                console.error(error.message);
+            }
         }
     }
